requirements.txt
asgiref==3.8.1
coverage==7.6.1
Django==5.1.1
django-debug-toolbar==4.4.6
django-taggit==6.0.0
Markdown==3.7
pillow==10.4.0
psycopg==3.2.2
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
